test(pairing): run invariant units once
